add config and strategy
