+++ schulcloud/Sync/AdditionalSyncData.swift
@@ -1,0 +1,15 @@
+//
+//  AdditionalSyncData.swift
+//  schulcloud
+//
+//  Created by Max Bothe on 09.02.18.
+//  Copyright © 2018 Hasso-Plattner-Institut. All rights reserved.
+//
+
+import Foundation
+
+struct AddtionalSyncData {
+
+    var externallyIncludedResourceData: [ResourceData]
+
+}
